Mild reorganization of package layout.
